requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
click==8.0.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.71.0
fastapi-mail==1.0.9
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
iso8601==1.0.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
Pillow==9.1.1
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
pyparsing==3.0.9
pypika-tortoise==0.1.5
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
redis==4.3.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
tortoise-orm==0.19.1
typing_extensions==4.0.1
uvicorn==0.16.0
wrapt==1.14.1
